pyproject.toml
[project]
name = "task-manager-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksandr",email = "94639744+ls500pymaster@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"

